STL files for the detector unit
There are 3 separate things to print. The mounting plate made for mounting on a 2020-extrusion, the housing itself containing the MQ135 unit, the LED, the NTC and the 6-pin JST connector and the NTC-mount.
Fix the housing on the mounting plate with 2 M3-bolts. Minimum length is 12mm. If you want to go to the end, you'll need a 20mm one and probably drill the bottom hole to 2.5mm.